// src/app/map/interfaces/marker.interface.ts
import { LngLat, Marker } from "mapbox-gl";


export interface MarkerColor{
    color : string,
    marker : Marker
}

export interface PlainMarker{
    color: string,
    lngLat : number[]
}

// Generated by https://quicktype.io

export interface Point {
    id:          number;
    name:        string;
    web:         string;
    description: string;
    adress:      string;
    lngLat:      [number, number];
    cat:         string[];
}